Update ReadME.txt with deepsparse-ent included (#699)
* Update ReadME.txt with deepsparse-ent included

Added in a proposed new ReadME for Deepsparse with refernces to the differences between deepsparse and deepsparse-ent as well as how to get started with enterprise. @bfineran and @markurtz plz review the section on deepsparse license getting set up with the env. variable. I was not sure of the detailed process here.

* Update README.md

* Quick fix formatting the folded enterprise info

* Update README.md

Co-authored-by: corey-nm <109536191+corey-nm@users.noreply.github.com>

* Updated links for MSLSA +trial; plus other changes

@bfineran still need review on line 135 to fill in the flow for env variable. Thanks

* added note on creating license.txt

* fixed Deepsparse to DeepSparse, plus Jeannie edit

* capitalization fix

* Final Version

This version should be final and once reviewed pushed to main. https://neuralmagic.com/legal/master-software-license-and-service-agreement/ this link will go live on Monday, 10/23/2022. All other links and wording should be final. 

@markurtz @bfineran please review and we'll look to push on Monday.

* removed word 'immediately' on trial description

Co-authored-by: corey-nm <109536191+corey-nm@users.noreply.github.com>